+++ tests/test_processing.py
@@ -258,3 +258,73 @@ class TestFilter(unittest.TestCase):
             plt.plot(x_axis, data_copy.data.data, label=filter_)
         plt.legend()
         plt.show()
+
+
+class TestTriggerAutodetection(unittest.TestCase):
+
+    def setUp(self):
+        self.dataset = trepr.dataset.ExperimentalDataset()
+        self.processing = trepr.processing.TriggerAutodetection()
+
+    def create_time_trace(self):
+        x1 = np.linspace(0, 10, 1000)
+        x2 = np.linspace(-2, 2, 200)
+        y1 = np.exp(-x1)
+        y1 = np.insert(y1, 0, np.zeros(100))
+
+        amplitude = 1
+        position = 0
+        width = 2
+        y2 = amplitude * np.exp(-(x2 - position) ** 2 / 2 * width ** 2)
+        convolved = np.convolve(y1, y2)
+        return convolved + np.random.random(len(convolved))
+
+    def test_instantiate_class(self):
+        pass
+
+    def test_has_appropriate_description(self):
+        self.assertIn('autodetect trigger position',
+                      self.processing.description.lower())
+
+    def test_with_1D_dataset_without_time_axis_raises(self):
+        dataset = trepr.dataset.ExperimentalDataset()
+        with self.assertRaises(aspecd.exceptions.NotApplicableToDatasetError):
+            dataset.process(self.processing)
+
+    def test_with_2D_dataset_without_time_axis_raises(self):
+        dataset = trepr.dataset.ExperimentalDataset()
+        dataset.data.data = np.random.random([5, 5])
+        with self.assertRaises(aspecd.exceptions.NotApplicableToDatasetError):
+            dataset.process(self.processing)
+
+    def test_with_2D_dataset_with_time_axis(self):
+        dataset = trepr.dataset.ExperimentalDataset()
+        dataset.data.data = np.random.random([5, 100])
+        dataset.data.axes[1].quantity = 'time'
+        dataset.process(self.processing)
+
+    def test_with_1D_dataset_sets_trigger_position(self):
+        self.dataset.data.data = self.create_time_trace()
+        self.dataset.data.axes[0].quantity = 'time'
+        self.dataset.process(self.processing)
+        self.assertGreaterEqual(
+            self.dataset.metadata.transient.trigger_position, 100)
+
+    def test_with_2D_dataset_sets_trigger_position(self):
+        self.dataset.data.data = np.tile(self.create_time_trace(), (5, 1))
+        self.dataset.data.axes[1].quantity = 'time'
+        self.dataset.process(self.processing)
+        self.assertGreaterEqual(
+            self.dataset.metadata.transient.trigger_position, 100)
+
+    def test_with_1D_dataset_sets_time_axis(self):
+        self.dataset.data.data = self.create_time_trace()
+        self.dataset.data.axes[0].quantity = 'time'
+        self.dataset.process(self.processing)
+        self.assertLess(self.dataset.data.axes[0].values[0], 0)
+
+    def test_with_2D_dataset_sets_time_axis(self):
+        self.dataset.data.data = np.tile(self.create_time_trace(), (5, 1))
+        self.dataset.data.axes[1].quantity = 'time'
+        self.dataset.process(self.processing)
+        self.assertLess(self.dataset.data.axes[1].values[0], 0)